fix: current number랑 questNumber 변경

+++ src/main/java/com/heartz/byeboo/adapter/in/web/dto/response/userquest/UserQuestDetailResponseDto.java
@@ -6,7 +6,7 @@ import com.heartz.byeboo.domain.model.UserQuest;
 import java.time.LocalDate;
 
 public record UserQuestDetailResponseDto(
-        Integer stepNumber,
+        Long stepNumber,
         Long questNumber,
         LocalDate createdAt,
         String question,
